+++ test_general/api/test_app.py
@@ -1,4 +1,5 @@
 from pathlib import Path
+import os
 
 import io
 import pytest
@@ -759,4 +760,30 @@ def test_general_api_returns_400_bad_docx():
         ],
     )
     assert "txt is not a valid" in response.json().get("detail")
+    assert response.status_code == 400
+
+
+def test_general_api_returns_400_bad_json(tmpdir):
+    """
+    Verify that we get a 400 for invalid json schemas
+    """
+    client = TestClient(app)
+    data = '{"hi": "there"}'
+
+    filepath = os.path.join(tmpdir, "unprocessable.json")
+    with open(filepath, "w") as f:
+        f.write(data)
+    response = client.post(
+        MAIN_API_ROUTE,
+        files=[
+            (
+                "files",
+                (
+                    str(filepath),
+                    open(filepath, "rb"),
+                ),
+            )
+        ],
+    )
+    assert "Unstructured schema" in response.json().get("detail")
     assert response.status_code == 400