+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 * Bump unstructured to 0.10.21
 * Fix an unhandled error when a non pdf file is sent with content-type pdf
-* Fix unhandled error when a non docx file is sent with content-type docx
+* Fix an unhandled error when a non docx file is sent with content-type docx
+* Fix an unhandled error when a non-Unstructured json schema is sent
 
 ## 0.0.51